requirements.txt
# OCR Project Requirements
# Updated for comprehensive OCR comparison project

# Core Python packages
packaging>=24.0
addict==2.4.0

# PyTorch with CUDA support for GPU acceleration
--find-links https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121

# PyTorch dependencies
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
pillow==11.0.0
setuptools==70.2.0
sympy==1.13.1
typing_extensions==4.12.2
transformers>=4.44.0

# Document processing
python-pptx==1.0.2
PyPDF2==3.0.0
lxml==6.0.0
xlsxwriter==3.2.5

# Windows-specific dependencies
comtypes==1.4.9
pywin32==306

# Legacy pathlib (for older Python versions)
pathlib==1.0.1

# OCR Libraries
# Surya OCR (install from source: git clone https://github.com/VikParuchuri/surya.git && cd surya && pip install -e .)
# PyTesseract (requires Tesseract binary installation)
pytesseract>=0.3.8
tesseract>=0.1.3

# Alternative OCR
easyocr>=1.7.1

# Image processing
opencv-contrib-python>=4.10.0.84

# Web framework (if needed for API)
flask>=2.0.2
flask-cors>=4.0.0

# PDF processing
pypdfium2>=4.0.0

# Ollama client
ollama>=0.1.7

# Additional utilities
requests>=2.31.0
urllib3>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: llama-cpp-python with CUDA support
# Uncomment the following lines if you want to install llama-cpp-python
# --extra-index-url https://jllllll.github.io/llama-cpp-python-cuBLAS-wheels/AVX2/cu121
# llama-cpp-python

# Installation Notes:
# 1. Create virtual environment: python -m venv ollama-ocr_env
# 2. Activate: .\ollama-ocr_env\Scripts\Activate.ps1
# 3. Install PyTorch first: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# 4. Install Surya from source: git clone https://github.com/VikParuchuri/surya.git && cd surya && pip install -e .
# 5. Install Tesseract binary separately (not via pip)
# 6. Install remaining requirements: pip install -r requirements.txt